requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
distro==1.8.0
Django==5.0
django-extensions==3.2.3
djangorestframework==3.15.2
et-xmlfile==1.1.0
exceptiongroup==1.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
numpy==2.0.0
openai==1.30.5
openpyxl==3.1.5
pandas==2.2.2
pandas-stubs==2.2.2.240603
psycopg==3.1.16
pydantic==2.5.2
pydantic_core==2.14.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2024.1
requests==2.32.3
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
types-pytz==2024.1.0.20240417
typing_extensions==4.9.0
tzdata==2024.1
urllib3==2.2.2
